usm: http2: Add path index
Adding to the current logic the path index, which will allow us to resolve the path (in
a future commit) in the user mode, rather than managing it in the kernel.

For the ROOT and INDEX paths, we're introducing constant (high) values, to mark them,
as both exists in the static table with a very low index, in which can cause us to mix
between new dynamic values with low index, and those paths.