usm: http2: Add interesting dynamic table hash map
The map acts as a set, to indicate if a given dynamic index (conn tuple + index) is interesting.
If a key exists - the index is interesting, otherwise it is not.

In a future commit we'll remove the old http2_dynamic_table map, and keep the new set instead
for kernel synchronization