usm: http2: Add user mode consumer to the perf buffer
The dynamic table management is now consuming the paths sent from the kernel.
We use an LRU datastore to automatically evict the least recently used element
when the map is full. On eviction we're deleting the matching entry from
http2_interesting_dynamic_table_set map.

The LRU value is a string internning pointer, to ensure we're using
a single instance and memory for the same path, even if the key is
different.